requirements.txt
diffusers==0.31.0
google_cloud_aiplatform==1.70.0
matplotlib==3.9.3
numpy==2.1.3
opencv_python==4.10.0.84
Pillow==11.0.0
protobuf==5.29.1
qrcode==8.0
Requests==2.32.3
st_clickable_images==0.0.3
streamlit==1.39.0
torch==2.5.1
